- decreased memory usage of sampler voices by sharing the temporary buffer

+++ hi_core/hi_sampler/sampler/components/SampleEditorComponents.cpp
@@ -799,7 +799,7 @@ void SamplerSoundMap::mouseDrag(const MouseEvent &e)
     refreshGraphics();
 }
 
-void SamplerSoundMap::setPressedKeys(const int *pressedKeyData)
+void SamplerSoundMap::setPressedKeys(const uint8 *pressedKeyData)
 {
 	for(int i = 0; i < 127; i++)
 	{
